+++ frontend2/src/components/compete/chart/chartUtils.ts
@@ -1,10 +1,37 @@
+import type { HistoricalRating, Tournament } from "api/_autogen";
 import type Highcharts from "highcharts";
 
-export const formatNumber = (value: number, decimals = 0): string =>
-  Number(value).toLocaleString(undefined, {
+type UTCMilliTimestamp = number;
+
+export type ChartData = [UTCMilliTimestamp, number];
+export type PlotLine = [string, UTCMilliTimestamp];
+
+export const formatNumber = (value: number, decimals = 0): string => {
+  return Number(value).toLocaleString(undefined, {
     minimumFractionDigits: decimals > 0 ? decimals : 0,
     maximumFractionDigits: decimals > 0 ? decimals : 0,
   });
+}
+
+export const formatRatingHistory = (
+  ratingHistory: HistoricalRating[],
+): Record<string, ChartData[]> | undefined => {
+  const ratingRecord: Record<string, ChartData[]> = {};
+  return ratingHistory.reduce((record, teamData) => {
+    if (teamData.team_rating !== undefined) {
+      record[teamData.team_rating.team.name] =
+        teamData.team_rating.rating_history.map((match) => [
+          match.timestamp.getTime(),
+          match.rating,
+        ]);
+    }
+    return record;
+  }, ratingRecord);
+};
+
+export const formatTournamentList = (tournaments: Tournament[]): PlotLine[] => {
+  return tournaments.map((t) => [t.name_long, t.display_date.getTime()]);
+};
 
 export const highchartsOptionsBase: Highcharts.Options = {
   chart: {
