+++ frontend2/src/views/Account.tsx
@@ -25,8 +25,8 @@ import {
 } from "../api/user/useUser";
 import { useEpisodeId } from "../contexts/EpisodeContext";
 import { useQueryClient } from "@tanstack/react-query";
-import type { QueryClient } from "@tanstack/query-core";
-import EpisodeChart from "components/compete/chart/EpisodeChart";
+import { type QueryClient } from "@tanstack/query-core";
+import UserChart from "components/compete/chart/UserChart";
 
 interface FileInput {
   file: FileList;
@@ -94,7 +94,7 @@ const Account: React.FC = () => {
         </SectionCard>
 
         <SectionCard title="Rating History" className="w-full flex-1">
-          <EpisodeChart />
+          <UserChart />
         </SectionCard>
       </div>
     </div>
